Add files via upload
Fixed an issue when creating the database with "Add CREATE DATABASE / USE statement", now you dont have to create the database in order to import the file, the file will do it automatically. I will still have to fix the problem with the event (for now, just import phplogin.sql and then, import the sqlevent file).